requirements.txt
Package             Version
------------------- ---------
asgiref             3.4.1
certifi             2021.5.30
charset-normalizer  2.0.6
Django              3.2.7
django-cors-headers 3.8.0
django-filter       2.4.0
django-truncate     0.1
djangorestframework 3.12.4
idna                3.2
Markdown            3.3.4
Pillow              8.3.2
pip                 21.2.4
pytz                2021.1
pyzbar              0.1.8
requests            2.26.0
setuptools          57.4.0
sqlparse            0.4.2
urllib3             1.26.7
wheel               0.37.0
